// src/reports/enums/report-type.enum.ts
export enum ReportType {
  USER_ACTIVITY = 'USER_ACTIVITY',
  TRANSACTION_SUMMARY = 'TRANSACTION_SUMMARY',
  PRODUCE_ANALYTICS = 'PRODUCE_ANALYTICS',
  QUALITY_METRICS = 'QUALITY_METRICS',
  MARKET_TRENDS = 'MARKET_TRENDS',
  FINANCIAL_SUMMARY = 'FINANCIAL_SUMMARY',
  INSPECTION_SUMMARY = 'INSPECTION_SUMMARY',
} 